Replace null fields for links with undefined
On entities we previously treated null fieldValues
as both a possible scalar or a hint that links should
be checked.

Instead we can use the undefined field and rely
on the presence of a selectionSet to see whether
a link should be read.

The field still needs to be set to undefined
however, so that the GC operation can walk
all entities.

We also have a special case for "invalid"
entities, which is a special case just in case
something goes very wrong.