Reimplement query operations using cascading nullables
First of all, this gets rid of `readOperation`, because
it's almost identical to `query`, which should save us
some bytes.

`readSelection` has been reimplemented. Instead of writing
fields onto data directly, it now first stores them in
`dataFieldValue`.

At any point when we encounter an uncached field that
is NOT nullable, we return undefined.
This is done by checking `dataFieldValue` against
the schema. This is now done in one place to save
some space and the hassle of reading repetitive code.

This works because we now represent uncached values
as `undefined`, which has been done in the past,
but has been removed. Hence:

- When `dataFieldValue` is `undefined` and a field
  is nullable then we write `null` and continue
- When `dataFieldValue` is `undefined` and a field
  is NOT nullable then we return `undefined` instead
  of `data` immediately
- Otherwise we write the field as usual

This is nice, because that means `undefined` cascades
upwards and this same logic also applies to resolved
links!