Fix operation being referred to but undefined
This is an interesting issue.
We thought it was impossible to be referring to the same
operation twice, since we're deduplicating for the operation
key when processing dependencies.
What we're not accounting for is that the operation may have
been reexecuted by another write to the cache.

This may happen in the following condition:

- A result comes back and a write operation is performed. Dependencies
  are stored as usual
- Another result comes back and triggers dependencies to be processed,
  which deletes some of the previous dependencies
- Meanwhile a potential third write starts which expects to see a
  deleted dependency, but it has already been reexecute by the second
  write operation.