Implement copy-first resolver result reading
This replaces the simple resolver result reader with
a "drilling" result reader, `readResolverResult` and
`resolveResolverResult`.

The former is similar to `readSelection` and the
latter is similar to `resolveLink`.

Instead of either resolving an entity or copying
the result of the resolver, we now read the data
from the resolver result if no key exists in the
store.