Add queryDependencies to updates in updateCacheWithResult
This is a significant change, where we may update more
queries than we have before. It adds queryDependencies
to the dependencies that are reexecuted when the cache is
updated, apart from just writeDependencies.

This is necessary because cache resolvers can alter
the data that is returned tremendously.

It should still be safe to do this though, because for
queries without any cache resolvers the queried data
should match the written data 1:1.

For data with cache resolvers that is queried however,
we will get some extra dependencies, which do need to
update.