Define new data structure
- records and links will be separated by (entity) node
- records and links will still have optimistic layers with partial entries
- every entity (by key) will have a reference count (RC)
- ref locks are entities with 0 ref count that are still kept for optimistic entries

Adding a link will increase the ref count (RC) by one.

When a link is modified we diff its entity keys with the previous link entry.
All deleted entity keys have their ref count (RC) decreased. If the RC
is zero we check whether it is referenced in an optimistic layer. If so the
entity key is put onto the ref lock stack by the optimistic key.

When that optimistic layer is removed we check all entity keys on the
ref lock stack. If a key on there still has an RC of zero it is deleted.