Remove case where scalar instead link is encountered
We remove this case because it's technically impossible
to encounter it. We can instead abort traversal at certain
points and not cache those values (readSelection modification).

We have a small identity function that will ensure that
TypeScript doesn't complain, which we can add warnings to
if we do want to check for this case.